DE2257: Specify an exact Date() on Observable.timer
Previously was specifying a timeout in milliseconds, but this was being used as an amount of time to wait before emitting, and that amount of time was "paused" when a device went to sleep.  Specifying a Date tells the timer to emit at that exact time (or after that time, if the time passes while a device is asleep).